usb: dwc3: W/A for dwc3 revision < 2.2a
WORKAROUND: For DWC3 revision < 2.20a have an issue
which would cause metastability state on Run/Stop
bit if we try to force the IP to USB2-only mode.

Because of that, we cannot configure the IP to any
speed other than the SuperSpeed

Refers to: STAR#9000525659: Clock Domain Crossing
on DCTL in USB 2.0 Mode.

Jira: None.
Test: Depends on patch "usb limit to high speed"
      Boot the board. Check adb with USB3.0 cable.

Change-Id: I5a40ac95990777626d3185b46ddef83bd4e66730
Signed-off-by: Abhilash K V <abhilash.k.v@intel.com>
Signed-off-by: karan889 <karanx.bhagoji@intel.com>